treadmill-rs: provide default handle_request impl for control_socket::Supervisor
